Put Cordova- and React Native-specific code into separate folders
