+++ lib/browser/fileReader.ts
@@ -1,4 +1,4 @@
-import { isReactNativeFile, isReactNativePlatform } from './isReactNative.js'
+import { isReactNativeFile, isReactNativePlatform } from '../reactnative/isReactNative.js'
 import { uriToBlob } from './uriToBlob.js'
 
 import type { FileReader, FileSource, UploadInput } from '../options.js'
