+++ lib/cordova/isCordova.ts
