+++ lib/cordova/readAsByteArray.ts
